move paths to gps/paths; remove FlattenOmitStdLib
