+++ cmd/dep/remove.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/golang/dep"
 	"github.com/golang/dep/internal/gps"
+	"github.com/golang/dep/internal/gps/paths"
 	"github.com/golang/dep/internal/gps/pkgtree"
-	"github.com/golang/dep/internal/paths"
 	"github.com/pkg/errors"
 )
 
@@ -72,7 +72,7 @@ func (cmd *removeCommand) Run(ctx *dep.Ctx, args []string) error {
 			return errors.Errorf("remove takes no arguments when running with -unused")
 		}
 
-		reachlist := reachmap.FlattenOmitStdLib()
+		reachlist := reachmap.FlattenFn(paths.IsStandardImportPath)
 
 		// warm the cache in parallel, in case any paths require go get metadata
 		// discovery
