+++ internal/gps/paths/paths.go
