+++ internal/gps/paths/paths_test.go
