+++ internal/gps/pkgtree/pkgtree_test.go
@@ -18,7 +18,7 @@ import (
 	"testing"
 
 	"github.com/golang/dep/internal/gps/internal/fs"
-	"github.com/golang/dep/internal/paths"
+	"github.com/golang/dep/internal/gps/paths"
 )
 
 // PackageTree.ToReachMap() uses an easily separable algorithm, wmToReach(),
