+++ internal/gps/solver.go
@@ -12,8 +12,8 @@ import (
 	"strings"
 
 	"github.com/armon/go-radix"
+	"github.com/golang/dep/internal/gps/paths"
 	"github.com/golang/dep/internal/gps/pkgtree"
-	"github.com/golang/dep/internal/paths"
 )
 
 var rootRev = Revision("")
